+++ cli/coder/input_guard.go
@@ -1,3 +1,9 @@
+/*
+ * ChatCLI - Command Line Interface for LLM interaction
+ * Copyright (c) 2024 Edilson Freitas
+ * License: Apache-2.0
+ */
+
 package coder
 
 import (
