+++ cli/coder/input_guard_linux.go
@@ -1,5 +1,11 @@
 //go:build linux
 
+/*
+ * ChatCLI - Command Line Interface for LLM interaction
+ * Copyright (c) 2024 Edilson Freitas
+ * License: Apache-2.0
+ */
+
 package coder
 
 import (
