+++ i18n/locales/en-US.json
@@ -885,8 +885,28 @@
   "agent.microcompact.applied": "microcompact: %d truncated, %d summarized, %s freed",
   "agent.queue.new_user_instruction": "New user instruction received",
   "agent.queue.indicator": "(%d queued)",
-  "slash.help.description": "Returns the chatcli command catalogue (slash and @-modifiers).",
+  "slash.help.description": "Returns the chatcli command catalog (slash and @-modifiers).",
   "slash.version.description": "Returns the running chatcli build version and whether an update is available.",
+  "plugins.websearch.describe": "Searching: %s",
+  "plugins.webfetch.describe": "Fetching: %s",
+  "plugins.coder.describe.read": "Reading: %s",
+  "plugins.coder.describe.search": "Searching: %s",
+  "plugins.coder.describe.exec": "Executing: %s",
+  "plugins.coder.describe.write": "Writing: %s",
+  "plugins.coder.describe.patch": "Patching: %s",
+  "plugins.coder.describe.tree": "Listing: %s",
+  "plugins.coder.describe.generic": "@coder %s",
+  "plugins.park.describe.delay": "Parking for: %s",
+  "plugins.park.describe.until": "Parking until: %s",
+  "plugins.park.describe.for_url": "Polling URL: %s",
+  "plugins.park.describe.for_cmd": "Polling cmd: %s",
+  "plugins.park.describe.generic": "@park %s",
+  "plugins.scheduler.describe.schedule": "Scheduling: %s",
+  "plugins.scheduler.describe.query": "@scheduler query: %s",
+  "plugins.scheduler.describe.cancel": "@scheduler cancel: %s",
+  "plugins.scheduler.describe.wait": "Waiting until: %s",
+  "plugins.scheduler.describe.list": "Listing scheduled jobs",
+  "plugins.scheduler.describe.generic": "@scheduler %s",
   "agent.early_exit.stagnation": "ReAct loop stopped: the same tool_call repeated %d turns in a row (stagnation). Set CHATCLI_AGENT_EARLY_EXIT=0 to disable this detector.",
   "agent.early_exit.chat_hint": "Tip: this query looks conversational — the /agent loop was skipped. Use /chat or just type the question to force chat mode, or /run to force agent mode.",
   "agent.webfetch.auto_saved": "web_fetch: response too large inline (%d bytes); saved to %s — use read_file or rerun with filter/from_line/to_line.",
