+++ i18n/locales/pt-BR.json
@@ -887,6 +887,26 @@
   "agent.queue.indicator": "(%d na fila)",
   "slash.help.description": "Retorna o catálogo de comandos do chatcli (slash e @-modificadores).",
   "slash.version.description": "Retorna a versão do chatcli em execução e se há atualização disponível.",
+  "plugins.websearch.describe": "Buscando: %s",
+  "plugins.webfetch.describe": "Baixando: %s",
+  "plugins.coder.describe.read": "Lendo: %s",
+  "plugins.coder.describe.search": "Buscando: %s",
+  "plugins.coder.describe.exec": "Executando: %s",
+  "plugins.coder.describe.write": "Escrevendo: %s",
+  "plugins.coder.describe.patch": "Patcheando: %s",
+  "plugins.coder.describe.tree": "Listando: %s",
+  "plugins.coder.describe.generic": "@coder %s",
+  "plugins.park.describe.delay": "Pausando por: %s",
+  "plugins.park.describe.until": "Pausando até: %s",
+  "plugins.park.describe.for_url": "Aguardando URL: %s",
+  "plugins.park.describe.for_cmd": "Aguardando cmd: %s",
+  "plugins.park.describe.generic": "@park %s",
+  "plugins.scheduler.describe.schedule": "Agendando: %s",
+  "plugins.scheduler.describe.query": "@scheduler query: %s",
+  "plugins.scheduler.describe.cancel": "@scheduler cancel: %s",
+  "plugins.scheduler.describe.wait": "Aguardando até: %s",
+  "plugins.scheduler.describe.list": "Listando tarefas agendadas",
+  "plugins.scheduler.describe.generic": "@scheduler %s",
   "agent.early_exit.stagnation": "loop ReAct interrompido: mesma tool_call repetida %d turnos seguidos (estagnação). Defina CHATCLI_AGENT_EARLY_EXIT=0 para desativar.",
   "agent.early_exit.chat_hint": "Dica: essa pergunta parece conversacional — o loop /agent foi ignorado. Use /chat ou digite a pergunta direto para forçar chat, ou /run para forçar o agente.",
   "agent.webfetch.auto_saved": "web_fetch: resposta grande demais para inlinear (%d bytes); salva em %s — use read_file ou repita com filter/from_line/to_line.",
